design: 삭제 버튼 레이아웃 완성
