+++ src/components/answer/AnswerPageCard.jsx
@@ -2,6 +2,7 @@ import styled from "styled-components";
 import { useState } from "react";
 import { getTimeAgo } from "utils/date";
 import AnswerComponent from "./AnswerComponent";
+import DeleteBtn from "./Delete";
 
 export function AnswerPageCard(props) {
   const { questionId, answer, content, question } = props;
